[upd] EditableProfileCard - moved ProfileCard from ProfilePage to EditableProfileCard
